Refactor users.ts into separate files to separate internal mutations, guards, queries, schema, seed functions, and helpers for user operations.
